backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
httpx==0.27.0
stripe==10.1.0
pyjwt==2.9.0
passlib==1.7.4
python-multipart==0.0.9
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
